# Cache tags from DbCommand