Parsing the connection string is a costly operation that we do at every query, even though we can reasonably expect users to have a limited set of connection strings. 

Benchmark:

```csharp
// With connection string: Server=myServerName,myPortNumber;Database=myDataBase;User Id=myUsername;Password=myPassword;
SpanExtensions.AddTagsFromDbCommand(span, dbCommand);

```

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method |       Runtime |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-------------- |-----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|    Old |    .NET 4.7.2 | 4,336.9 ns | 49.97 ns | 44.30 ns |  1.00 | 0.3433 |     - |     - |    2182 B |
|    New |    .NET 4.7.2 |   242.4 ns |  1.54 ns |  1.45 ns |  0.06 |      - |     - |     - |         - |
|        |               |            |          |          |       |        |       |       |           |
|    Old | .NET Core 3.1 | 3,495.5 ns | 12.84 ns | 10.02 ns |  1.00 | 0.3510 |     - |     - |    2208 B |
|    New | .NET Core 3.1 |   236.4 ns |  3.97 ns |  3.52 ns |  0.07 |      - |     - |     - |         - |