@kevingosse 2020-06-30
2 things to take care of before merging:
- This would cause a memory leak if there's some random part in a connection string. I can't think of a situation where that would happen, but I'm going to do some research to make sure
- I'll add a unit test