@kevingosse 2020-07-01
I could think of a situation where we would have an infinite number of connection strings: if the user opens arbitrary Excel files with ADO.NET. This is very unlikely but still possible, so I added a fail-safe mechanism. There's no noticeable impact on the benchmark (expected, since the fast path doesn't change).

``` ini

BenchmarkDotNet=v0.12.0, OS=Windows 10.0.19041
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-VTGAFA : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  Job-MNUBEA : .NET Core 3.1.5 (CoreCLR 4.700.20.26901, CoreFX 4.700.20.27001), X64 RyuJIT


```
| Method |       Runtime |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |-------------- |-----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|    Old |    .NET 4.7.2 | 4,464.6 ns | 72.36 ns | 67.69 ns |  1.00 | 0.3433 |     - |     - |    2182 B |
|    New |    .NET 4.7.2 |   244.2 ns |  4.59 ns |  4.50 ns |  0.05 |      - |     - |     - |         - |
|        |               |            |          |          |       |        |       |       |           |
|    Old | .NET Core 3.1 | 3,711.8 ns | 73.20 ns | 78.33 ns |  1.00 | 0.3510 |     - |     - |    2208 B |
|    New | .NET Core 3.1 |   242.2 ns |  2.32 ns |  1.94 ns |  0.07 |      - |     - |     - |         - |